+++ test/test_examples.py
@@ -87,10 +87,11 @@ def test_passthrough(tmpdir, name, debug, capfd):
         tst_readdir(src_dir, work_dir)
         tst_open_read(src_dir, work_dir)
         tst_open_write(src_dir, work_dir)
+        tst_create(work_dir)
+        tst_passthrough(src_dir, work_dir)
         if not is_ll:
             tst_mkdir(work_dir)
             tst_rmdir(src_dir, work_dir)
-            tst_create(work_dir)
             tst_unlink(src_dir, work_dir)
             tst_symlink(work_dir)
             if os.getuid() == 0:
@@ -103,7 +104,6 @@ def test_passthrough(tmpdir, name, debug, capfd):
             tst_truncate_path(work_dir)
             tst_truncate_fd(work_dir)
             tst_open_unlink(work_dir)
-            tst_passthrough(src_dir, work_dir)
 
             subprocess.check_call([ os.path.join(basename, 'test', 'test_syscalls'),
                                     work_dir, ':' + src_dir ])