+++ example/cuse.c
@@ -31,7 +31,7 @@
  */
 
 
-#define FUSE_USE_VERSION 30
+#define FUSE_USE_VERSION 31
 
 #include <config.h>