+++ example/hello.c
@@ -19,7 +19,7 @@
  */
 
 
-#define FUSE_USE_VERSION 30
+#define FUSE_USE_VERSION 31
 
 #include <config.h>