+++ example/ioctl.c
@@ -22,7 +22,7 @@
  * \include ioctl.c
  */
 
-#define FUSE_USE_VERSION 30
+#define FUSE_USE_VERSION 31
 
 #include <config.h>