+++ example/notify_inval_entry.c
@@ -73,7 +73,7 @@
  */
 
 
-#define FUSE_USE_VERSION 30
+#define FUSE_USE_VERSION 31
 
 #include <config.h>