+++ example/notify_inval_inode.c
@@ -59,7 +59,7 @@
  */
 
 
-#define FUSE_USE_VERSION 30
+#define FUSE_USE_VERSION 31
 
 #include <config.h>