+++ example/passthrough_ll.c
@@ -28,7 +28,7 @@
  */
 
 #define _GNU_SOURCE
-#define FUSE_USE_VERSION 30
+#define FUSE_USE_VERSION 31
 
 #ifdef HAVE_CONFIG_H
 #include <config.h>