+++ ChangeLog.rst
@@ -3,6 +3,10 @@ Unreleased Changes
 
 * `fuse_main()` / `fuse_remove_signal_handlers()`: do not reset
   `SIGPIPE` handler to `SIG_DFL` it was not set by us.
+* Documented the `RENAME_EXCHANGE` and `RENAME_NOREPLACE` flags that
+  may be passed to the `rename` handler of both the high- and
+  low-level API. Filesystem authors are strongly encouraged to check
+  that these flags are handled correctly.
 
 libfuse 3.0.2 (2017-05-24)
 ==========================