+++ ChangeLog.rst
@@ -6,7 +6,7 @@ Unreleased Changes
 * The `example/passthrough_ll` filesystem now supports creating
   and writing to files.
 * `fuse_main()` / `fuse_remove_signal_handlers()`: do not reset
-  `SIGPIPE` handler to `SIG_DFL` it was not set by us.
+  `SIGPIPE` handler to `SIG_DFL` if it was not set by us.
 * Documented the `RENAME_EXCHANGE` and `RENAME_NOREPLACE` flags that
   may be passed to the `rename` handler of both the high- and
   low-level API. Filesystem authors are strongly encouraged to check